Agregado objeto SOLIDO para colisiones

+++ src/server_src/objeto.cpp
@@ -2,7 +2,7 @@
 
 
 
-Objeto::Objeto(uint32_t id_objeto, uint32_t posicion_x, uint32_t posicion_y, double tiempo_reaparicion, uint8_t ancho, uint8_t alto) :
+Objeto::Objeto(uint32_t id_objeto, uint32_t posicion_x, uint32_t posicion_y, double tiempo_reaparicion, uint32_t ancho, uint32_t alto) :
         id_objeto(id_objeto), posicion_x(posicion_x), posicion_y(posicion_y),
         tiempo_reaparicion(tiempo_reaparicion), tiempo_interaccion(-tiempo_reaparicion),
         mostrar(true), ancho(ancho), alto(alto)
@@ -18,11 +18,11 @@ uint32_t Objeto::obtener_posicionY(){
     return posicion_y;
 }
 
-uint8_t Objeto::obtener_ancho(){
+uint32_t Objeto::obtener_ancho(){
     return ancho;
 }
 
-uint8_t Objeto::obtener_alto(){
+uint32_t Objeto::obtener_alto(){
     return alto;
 }
